refactor: separate share intent to the utils file
