+++ app/src/main/java/com/imeanttobe/consensusapp/core/Utils.kt
@@ -3,6 +3,7 @@ package com.imeanttobe.consensusapp.core
 import android.app.Activity
 import android.content.Context
 import android.content.ContextWrapper
+import android.content.Intent
 
 fun Context.findActivity(): Activity? {
     var context = this
@@ -13,4 +14,14 @@ fun Context.findActivity(): Activity? {
         context = context.baseContext
     }
     return null
+}
+
+fun openShareIntent(title: String, content: String, context: Context) {
+    val sendIntent = Intent().apply {
+        action = Intent.ACTION_SEND
+        putExtra(Intent.EXTRA_TEXT, content)
+        type = "text/plain"
+    }
+    val shareIntent = Intent.createChooser(sendIntent, title)
+    context.startActivity(shareIntent)
 }
